third_party/blink/web_tests/flag-specific/highdpi/virtual/container-queries/wpt_internal/css/css-conditional/container-queries/at-container-003-expected.txt

This is a testharness.js-based test.
FAIL Children respond to changes in container size assert_equals: expected "rgb(255, 0, 0)" but got "rgb(0, 128, 0)"
FAIL Descendants respond to changes in container size assert_equals: expected "rgb(255, 0, 0)" but got "rgb(0, 128, 0)"
Harness: the test ran to completion.
